requirements.txt
beautifulsoup4>=4.10.0
pandas>=1.5.1
psycopg2-binary>=2.9.5